# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      description: index hello
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: index test
      tags:
      - index
  /user/createUser:
    post:
      parameters:
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: repassword
        in: formData
        name: repassword
        type: string
      responses:
        "200":
          description: code","msg"}
          schema:
            type: string
      summary: 新增用户
      tags:
      - 用户模块
  /user/deleteUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      responses:
        "200":
          description: code","msg"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/getUserList:
    get:
      responses:
        "200":
          description: code","data"}
          schema:
            type: string
      summary: 查询所有用户
      tags:
      - 用户模块
  /user/login:
    post:
      parameters:
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: code","msg","data"}
          schema:
            type: string
      summary: 登录
      tags:
      - 用户模块
  /user/updateUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 更新用户
      tags:
      - 用户模块
swagger: "2.0"